solved finally

+++ codechef/may_challenge_2018_div2/dibs_on_fibs_DBFB/dibs_on_fibs.cpp
@@ -32,7 +32,7 @@ int main() {
         //do fibonacci using those sums as initial numbers
         ll result = 0;
         int k;
-        vector<int> fib;
+        vector<ll> fib;
         fib.resize(max(2,N));
         fib[0] = sum1;
         fib[1] = sum2;
